Added bot run and token code
With this you now have to set your bot token as an environment variable, e.g., `export JARVIS_TOKEN="<token>"`, and when you launch the application the token is passed in and the application connects. This will be useful when we set up automatic deployment and/or testing further down the road.