Remove link preview for articles
Due to the fact that a long list of articles (or even a fairly short
list) can create a lot of space in the form of link previews, this
commit removes those previews in all cases. Achieved by placing the
link in angle brackets ('<' and '>'), which disables link preview on the
Discord end.